Module aws_sdk_codeguruprofiler::error

Modules

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AddNotificationChannelsError

BatchGetFrameMetricDataError

ConfigureAgentError

ConflictException

The requested operation would cause a conflict with the current state

CreateProfilingGroupError

DeleteProfilingGroupError

DescribeProfilingGroupError

GetFindingsReportAccountSummaryError

GetNotificationConfigurationError

GetPolicyError

GetProfileError

GetRecommendationsError

InternalServerException

The server encountered an internal error and is unable to complete the request.

ListFindingsReportsError

ListProfileTimesError

ListProfilingGroupsError

ListTagsForResourceError

PostAgentProfileError

PutPermissionError

RemoveNotificationChannelError

RemovePermissionError

ResourceNotFoundException

The resource specified in the request does not exist.

ServiceQuotaExceededException

You have exceeded your service quota. To perform the requested action,

SubmitFeedbackError

TagResourceError

ThrottlingException

The request was denied due to request throttling.

UntagResourceError

UpdateProfilingGroupError

ValidationException

The parameter is not valid.

Enums

AddNotificationChannelsErrorKind

BatchGetFrameMetricDataErrorKind

ConfigureAgentErrorKind

CreateProfilingGroupErrorKind

DeleteProfilingGroupErrorKind

DescribeProfilingGroupErrorKind

GetFindingsReportAccountSummaryErrorKind

GetNotificationConfigurationErrorKind

GetPolicyErrorKind

GetProfileErrorKind

GetRecommendationsErrorKind

ListFindingsReportsErrorKind

ListProfileTimesErrorKind

ListProfilingGroupsErrorKind

ListTagsForResourceErrorKind

PostAgentProfileErrorKind

PutPermissionErrorKind

RemoveNotificationChannelErrorKind

RemovePermissionErrorKind

SubmitFeedbackErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateProfilingGroupErrorKind