Module aws_sdk_codeguruprofiler::model

Modules

agent_configuration

See AgentConfiguration

agent_orchestration_config

See AgentOrchestrationConfig

aggregated_profile_time

See AggregatedProfileTime

anomaly

See Anomaly

anomaly_instance

See AnomalyInstance

channel

See Channel

findings_report_summary

See FindingsReportSummary

frame_metric

See FrameMetric

frame_metric_datum

See FrameMetricDatum

match

See Match

metric

See Metric

notification_configuration

See NotificationConfiguration

pattern

See Pattern

profile_time

See ProfileTime

profiling_group_description

See ProfilingGroupDescription

profiling_status

See ProfilingStatus

recommendation

See Recommendation

timestamp_structure

See TimestampStructure

user_feedback

See UserFeedback

Structs

AgentConfiguration

AgentOrchestrationConfig

AggregatedProfileTime

Anomaly

AnomalyInstance

The specific duration in which the metric is flagged as anomalous.

Channel

Notification medium for users to get alerted for events that occur in application profile. We support SNS topic as a notification channel.

FindingsReportSummary

FrameMetric

FrameMetricDatum

Match

The part of a profile that contains a recommendation found during analysis.

Metric

NotificationConfiguration

The configuration for notifications stored for each profiling group. This includes up to

Pattern

ProfileTime

ProfilingGroupDescription

ProfilingStatus

Recommendation

A potential improvement that was found from analyzing the profiling data.

TimestampStructure

UserFeedback

Feedback that can be submitted for each instance of an anomaly by the user.

Enums

ActionGroup

AgentParameterField

AggregationPeriod

ComputePlatform

EventPublisher

FeedbackType

MetadataField

MetricType

OrderBy